requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-sendgrid-webhook==0.3.0
djangorestframework==3.14.0
idna==3.4
lxml==4.9.2
numpy==1.24.3
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3
requests==2.28.2
sendgrid==6.10.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tzdata==2023.3
urllib3==1.26.15
